requirements.txt
numpy==2.3.5
pillow==12.0.0
pyglm==2.8.3
